Improve removal process in image services
Changes have been made to handle the removal of album and image data more effectively. These modifications not only ensure to retrieve all necessary data before deletion but also handle error cases better.

src/modules/images/services/album.service.ts:
+ Added an image removal loop for each image of an album before its deletion
+ Added removal for album cover and album cover thumbnail files if they exist

src/modules/firebase/services/storage.service.ts:
+ Added error handling in the file removal process; specifically handling 'not found' and 'unauthorized' cases

src/modules/images/entities/album.entity.ts:
+ Made cover and cover thumbnail eager loading when querying album entity

src/modules/images/services/image.service.ts:
+ Refactored file removal calls into a new method 'removeFromEntity'

+++ src/modules/firebase/services/storage.service.ts
@@ -75,12 +75,19 @@ export class StorageService {
   }
 
   async removeFile(path: string): Promise<void> {
-    if (!this._storage)
-      this._storage = getStorage().bucket(this._cs.get('firebase.storage.bucket'));
+    try {
+      if (!this._storage)
+        this._storage = getStorage().bucket(this._cs.get('firebase.storage.bucket'));
 
-    const file = this._storage.file(path);
+      const file = this._storage.file(path);
 
-    await file.delete();
+      if (!file) return;
+
+      await file.delete();
+    } catch (error) {
+      if (error.code === 404) return;
+      if (error.code === 403) throw new Error('Unauthorized to remove file');
+    }
   }
 
   async getDownloadUrl(path: string): Promise<string> {
